at least it compiles, still removing pointless templating

+++ lpc/clocks.h
@@ -10,12 +10,11 @@ unsigned pllInputHz(bool forUsb);
 
 /** @returns the frequency pre-divider */
 unsigned coreInputHz();
-/** @returns the frequency post-divider
+/** @returns the approx frequency post-divider, rounded to nearest Hz
  * even if xtal is in integer Hz there is a divisor that isn't a power of 2 */
-double coreHz(void);
+unsigned coreHz(void);
 
-
-/** return Hz for @param bus number, -1 for systick, 0 for core, 1.. for AHB etc device specific values usually tied to apb index/*/
+/** @returns Hz for @param bus number, -1 for systick, 0 for core, 1.. for AHB etc device specific values usually tied to apb index/*/
 unsigned clockRate(int bus);
 
 /** set clock to fastest possible with @internal osc else with its own logic as to which is faster.*/
